requirements.txt
tikhub>=2.1.1
google-genai>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
snscrape>=0.7.0
flask>=3.0.0


